Moved code to a detail header file.
